--- src/com/company/tasks/Task2.java
@@ -1,5 +1,0 @@
-package com.company;
-
-public class Task2 {
-    
-}
