--- src/com/company/tasks/Task3.java
@@ -1,7 +1,0 @@
-package com.company.tasks;
-
-//03_chapter_V_3
-public class Task3 {
-
-
-}
