--- src/com/company/tasks/Task4.java
@@ -1,2 +1,0 @@
-package com.company;public class Task4 {
-}
