--- src/com/company/tasks/Task5.java
@@ -1,2 +1,0 @@
-package com.company;public class Task5 {
-}
